+++ src/main/java/site/siredvin/progressiveperipherals/integrations/computercraft/turtles/BrewingTurtle.java
@@ -1,0 +1,35 @@
+package site.siredvin.progressiveperipherals.integrations.computercraft.turtles;
+
+import dan200.computercraft.api.turtle.ITurtleAccess;
+import dan200.computercraft.api.turtle.TurtleSide;
+import de.srendi.advancedperipherals.common.addons.computercraft.base.ModelTransformingTurtle;
+import net.minecraft.client.renderer.model.ModelResourceLocation;
+import net.minecraft.item.ItemStack;
+import net.minecraft.util.ResourceLocation;
+import org.jetbrains.annotations.NotNull;
+import site.siredvin.progressiveperipherals.ProgressivePeripherals;
+import site.siredvin.progressiveperipherals.common.setup.Items;
+import site.siredvin.progressiveperipherals.integrations.computercraft.peripherals.automata.BrewingAutomataCorePeripheral;
+import site.siredvin.progressiveperipherals.utils.TranslationUtil;
+
+public class BrewingTurtle extends ModelTransformingTurtle<BrewingAutomataCorePeripheral> {
+    public static final ResourceLocation ID = new ResourceLocation(ProgressivePeripherals.MOD_ID, "brewing_automata");
+    public BrewingTurtle() {
+        super(ID, TranslationUtil.turtle("brewing"), new ItemStack(Items.BREWING_AUTOMATA_CORE.get()));
+    }
+
+    @Override
+    protected ModelResourceLocation getLeftModel() {
+        return null;
+    }
+
+    @Override
+    protected ModelResourceLocation getRightModel() {
+        return null;
+    }
+
+    @Override
+    protected BrewingAutomataCorePeripheral buildPeripheral(@NotNull ITurtleAccess turtle, @NotNull TurtleSide side) {
+        return new BrewingAutomataCorePeripheral("brewingAutomataCore", turtle, side);
+    }
+}